orderbook logic is added for sell and buy a commodity

+++ exchange/common/src/main/java/bist/demo/exchange/common/message/Command.java
@@ -1,20 +1,20 @@
 package bist.demo.exchange.common.message;
 
 public enum Command {
-    HEARTBEAT(0x48),
-    SEND_ORDER( 0x4F);
+    HEARTBEAT((byte) 0x48),
+    SEND_ORDER((byte) 0x4F);
 
-    private final int value;
+    private final byte value;
 
-    Command(int value) {
+    Command(byte value) {
         this.value = value;
     }
 
-    public int getValue(){
+    public byte getValue(){
         return this.value;
     }
 
-    public static Command isValidCommand(int value){
+    public static Command isValidCommand(byte value){
         for(Command command : Command.values()){
             if(value == command.getValue()) return command;
         }
